The partner application is reviewed and backend connected successfully
